Split graph page to components
