+++ src/components/Graph/SvgText.js
@@ -4,27 +4,29 @@ import { intensity, colorPalette } from "../../store/rpeStrings";
 
 function SvgText(props) {
   const { mass, reps, sets, rpe, date } = props.showValue;
-  
+
   return (
     <>
-      <circle
-        id="3"
-        cx="12"
-        cy="50%"
-        r="8"
-        stroke="rgb(0, 0, 0, 20%)"
-        strokeWidth="2"
-        fill={colorPalette[rpe]}
-      ></circle>
-      <text className={css.graphText} x={25} y="30%">
-        {date.day}.{date.month}.{date.year}
-      </text>
-      <text className={css.graphText} x={25} y="60%">
-        {mass}kg, {reps} reps for {sets} sets
-      </text>
-      <text className={css.graphText} x={25} y="90%">
-        {intensity[rpe]}
-      </text>
+      <svg height="50" width="30">
+        <circle
+          id="3"
+          cx="50%"
+          cy="50%"
+          r="8"
+          stroke="rgb(0, 0, 0, 20%)"
+          strokeWidth="2"
+          fill={colorPalette[rpe]}
+        ></circle>
+      </svg>
+      <div>
+        <p className={css.graphText}>
+          {date.day}.{date.month}.{date.year}
+        </p>
+        <p className={css.graphText}>
+          {mass}kg, {reps} reps for {sets} sets
+        </p>
+        <p className={css.graphText}>{intensity[rpe]}</p>
+      </div>
     </>
   );
 }
